backend/requirements.txt
fastapi
uvicorn
python-multipart
llama-parse
anthropic
python-dotenv
pandas
